refactor(game): #22 stop vf attack pattern reuse
There's still a lot of assumptions to be made that won't apply to ships
which don't use the ammunition system, let alone the weapon, flight and
AI behaviors used by the missile corvette and the missile destroyer.

A bare minimum API should allow at least some degree of flexibility for
ships that will use a different set of conditions (e.g. energy pool) to
perform a volley volley fire attack.